refactor: extract SwapFormUiModel and sub-models into SwapFormUiModel.kt (#4775)
Move QuoteDisplay, FeeBreakdown, DiscountInfo and SwapFormUiModel out of
SwapFormViewModel.kt into the dedicated SwapFormUiModel.kt write target
listed in the issue, dropping the VM from 1197 to 1145 lines (below the
1185 base) to deliver the headline shrink goal.
